Fix the makefile override in tests.yaml
Config merges it’s defaults against those in tests.yaml. In the case of
the makefile setting, it appends to the makefile targets to run, rather
than replacing. This can cause duplicate tests to be run and tests run
that shouldn’t be. This fix call’s update, which will replace the key
contents.